+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added TLS support to native client
 - Bumped minimum supported Rust version to `1.51`
 - Pinned `wasm-bindgen` dependency to `0.2.74`